requirements.txt
streamlit>=1.28.0
openai>=2.15.0
python-dotenv>=1.0.0
# PyMuPDF (fitz) - Best for Arabic PDFs with RTL support
PyMuPDF>=1.23.0
# Keep PyPDF2 as fallback
pypdf2>=3.0.0
pdf2image>=1.16.3
pillow>=10.0.0
# Tesseract OCR with Arabic support
pytesseract>=0.3.10
# EasyOCR for low-quality Arabic images
easyocr>=1.7.0
pydantic>=2.5.0
python-dateutil>=2.8.2
